traffic/requirements.txt
aiohttp
SQLAlchemy
psycopg2-binary
python-dotenv
psutil
xgboost
numpy
pandas
joblib
scikit-learn